tidy + move logic to fn
