+++ compiler/rustc_error_messages/locales/en-US/hir_analysis.ftl
@@ -147,6 +147,6 @@ hir_analysis_const_impl_for_non_const_trait =
 hir_analysis_const_bound_for_non_const_trait =
     ~const can only be applied to `#[const_trait]` traits
 
-hir_analysis_self_in_impl_self = 
+hir_analysis_self_in_impl_self =
     `Self` is not valid in the self type of an impl block
     .note = replace `Self` with a different type
